// src/app/shared/type.ts
export type Specialist = {
    id?: string;
    img?: string;
    name?: string;
    detail?: string;
    position?: string;
}
  
export type Banner = {
    id?: string;
    img?: string;
    description?: string;
    date?: string;
    redirectTo?: string;
}

export type Activity = {
    id?: string;
    img?: string;
    title?: string;
    description?: string;
    date?: string;
    redirectTo?: string;
}

export type ServiceUnit = {
    id?: string;
    img?: string;
    name?: string;
    detail?: string;
}